Move SequenceTypes to enums file
