+++ src/enums/sequencing.ts
@@ -1,0 +1,4 @@
+export enum SequenceTypes {
+  LIBRARY = 'library',
+  USER = 'user',
+}
